Bug fixes: TICK TOCK Added: stored error mgs
The time calc did not take the minutes rolling over a new hour, plus $(()) did not
like leading 0 on any number.  A better solution is to store TICK and TOCK as seconds since the epoch
which makes the subtraction straightforward and removes the leading 0 problem.

Added a variable to display an error msgs from wget and display them at completion.